Not necessary to include namespace in kubectl apply command
Should be defined in the yaml directly. All traefik services,
ingresses and the ingress controller should be in the kube-system
namespace. We put traefik stuff in the kube-system namespace to
make a distinction between things that run on the cluster and
things that are components of the cluster. It is purely organizational.